requirements.txt
PyMuPDF
pytesseract
opencv_python
pytest
Pillow